Produce every artifact through Linker; fold Bundler into the scheme
A new engine-independent `anthology.linker` module hosts the
`Linker`/`Linkage` machinery (kept out of capture-checked `core`, which
rejects the option-capture and zip-traversal idioms) plus JAR packaging:
`Linkage[Artifact.Jar]` builds executable JARs (at most one entry point,
`jarOptions.name` choosing the filename) and
`Linkage[Artifact.Library[universe]]`, available in every universe, packages
unlinked output for downstream assembly — both zero-import, living in
`Linkage`'s companion. The typed `Bundler.bundle`/`library` methods are gone;
`Bundler` retains only the classloader-based self-replication path the
staging rigs use, which no longer drags linker engines onto their classpaths.
The sjsir linkages move to an `import sjsLinkages.given` import object, the
one ceremony this reshuffle costs.
