+++ lib/anthology/src/linker/anthology.LinkError.scala
@@ -30,8 +30,21 @@
 ┃                                                                                                  ┃
 ┗━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━┛
                                                                                                   */
-package soundness
+package anthology
 
-export
-  anthology
-  . { linkerOptions, sjsLinkages, WasiToolchain, WitWorld }
+import digression.*
+import fulminate.*
+
+object LinkError:
+  enum Reason(val number: Int) extends Clarification:
+    case Failed(trace: StackTrace) extends Reason(1)
+    case NoEntryPoint              extends Reason(3)
+    case ManyEntryPoints           extends Reason(4)
+
+  given communicable: Reason is Communicable =
+    case Reason.Failed(_)       => m"the linker terminated abnormally"
+    case Reason.NoEntryPoint    => m"a native executable requires exactly one entry point"
+    case Reason.ManyEntryPoints => m"an executable JAR permits at most one entry point"
+
+case class LinkError(reason: LinkError.Reason)(using Diagnostics)
+extends Error(779, reason.number)(m"linking failed because $reason")
