+++ lib/anthology/src/linker/anthology.Linkage.scala
@@ -32,17 +32,86 @@
                                                                                                   */
 package anthology
 
+import java.nio.file as jnf
+
+import scala.util.control as suc
+
+import anticipation.*
+import contingency.*
 import digression.*
-import fulminate.*
+import galilei.*
+import hellenism.*
+import prepositional.*
+import serpentine.*
+import vacuous.*
 
-object LinkError:
-  enum Reason(val number: Int) extends Clarification:
-    case Failed(trace: StackTrace) extends Reason(1)
-    case NoEntryPoint              extends Reason(3)
+object Linkage:
+  // The packaging link family: JAR files, assembled by `Bundler` from a compilation's output.
+  // `Form` is the JAR's filename within the output directory.
+  given jar: (Linkage[Artifact.Jar] from Universe.Classfile):
+    type Origin = Universe.Classfile
+    private[anthology] type Form = Text
+    private[anthology] def initial: Text = "main.jar".tt
 
-  given communicable: Reason is Communicable =
-    case Reason.Failed(_)    => m"the linker terminated abnormally"
-    case Reason.NoEntryPoint => m"a native executable requires exactly one entry point"
+    private[anthology] def link
+      ( form:        Text,
+        compilation: Compilation[Universe.Classfile],
+        entryPoints: List[Linker.EntryPoint],
+        out:         Path on Linux )
+    :   Path on Linux logs LinkEvent raises LinkError =
 
-case class LinkError(reason: LinkError.Reason)(using Diagnostics)
-extends Error(779, reason.number)(m"linking failed because $reason")
+      val main: Optional[Fqcn] = entryPoints match
+        case Nil         => Unset
+        case List(entry) => entry.mainClass
+        case _           => abort(LinkError(LinkError.Reason.ManyEntryPoints))
+
+      try
+        jnf.Files.createDirectories(jnf.Paths.get(out.encode.s))
+
+        val entries =
+          Classpath.Directory(compilation.out) :: compilation.classpath.entries.to(List)
+
+        unsafely(Bundler.assemble(LocalClasspath(entries*), out / form, main))
+
+      catch case suc.NonFatal(error) =>
+        abort(LinkError(LinkError.Reason.Failed(error.stackTrace)))
+
+  // A library JAR is available in every universe, packaging whatever the compilation
+  // emitted—classfiles, and any `.sjsir` or `.nir` alongside them—for downstream assembly.
+  // Entry points do not apply and are ignored.
+  given library: [universe <: Universe] => (Linkage[Artifact.Library[universe]] from universe):
+    type Origin = universe
+    private[anthology] type Form = Text
+    private[anthology] def initial: Text = "main.jar".tt
+
+    private[anthology] def link
+      ( form:        Text,
+        compilation: Compilation[universe],
+        entryPoints: List[Linker.EntryPoint],
+        out:         Path on Linux )
+    :   Path on Linux logs LinkEvent raises LinkError =
+
+      try
+        jnf.Files.createDirectories(jnf.Paths.get(out.encode.s))
+        val entries = List(Classpath.Directory(compilation.out))
+        unsafely(Bundler.assemble(LocalClasspath(entries*), out / form, Unset))
+
+      catch case suc.NonFatal(error) =>
+        abort(LinkError(LinkError.Reason.Failed(error.stackTrace)))
+
+// Determines how an artifact is linked from a compilation in its origin universe: the
+// underlying linker-configuration type (`Form`), the configuration the artifact mandates, and
+// the link pipeline itself. Instances exist only for artifacts that are currently producible,
+// and those whose prerequisites can be absent are conditional upon evidence of them: the WASI
+// instance (in the `link` module) upon a `WasiToolchain` and a `WitWorld`, and the
+// native-binary instance (in the `nir` module) upon probing the C toolchain.
+trait Linkage[artifact <: Artifact] extends Provenance[artifact]:
+  private[anthology] type Form
+  private[anthology] def initial: Form
+
+  private[anthology] def link
+    ( form:        Form,
+      compilation: Compilation[Origin],
+      entryPoints: List[Linker.EntryPoint],
+      out:         Path on Linux )
+  :   Path on Linux logs LinkEvent raises LinkError
