+++ lib/anthology/src/linker/anthology_linker.scala
@@ -32,110 +32,9 @@
                                                                                                   */
 package anthology
 
-import ambience.*
 import anticipation.*
-import contingency.*
-import digression.*
-import distillate.*
-import eucalyptus.*
-import galilei.*
-import gossamer.*
-import hellenism.*
-import prepositional.*
-import revolution.*
-import serpentine.*
-import turbulence.*
-import vacuous.*
-import zeppelin.*
 
-import filesystemBackends.virtualMachine
-import filesystemOptions.dereferenceSymlinks.enabled
-import filesystemTraversal.preOrderTraversal
-import logging.silentLogging
-import manifestAttributes.*
-import systems.javaSystem
-import workingDirectories.javaWorkingDirectory
-
-object Bundler:
-  def classpath(out: Path on Linux): LocalClasspath =
-    val entries = Classpath.Directory(out) :: (classloaders.threadContextClassloader.classpath.match
-      case classpath: LocalClasspath => classpath.entries
-
-      case _ =>
-        unsafely(System.properties.java.`class`.path().as[LocalClasspath]).entries)
-
-    LocalClasspath(entries*)
-
-
-  def bundle(directory: Path on Linux, jarfile0: Optional[Path on Linux], main: Optional[Fqcn])
-  :   Path on Linux raises ZipError raises PathError raises IoError raises StreamError =
-
-    assemble(classpath(directory), jarfile0.or(directory.peer("tmpfile.jar")), main)
-
-
-  // Bundles a classfile compilation, together with the classpath it was compiled against, as an
-  // executable JAR file.
-  def bundle
-    ( compilation: Compilation[Universe.Classfile],
-      jarfile0:    Optional[Path on Linux],
-      main:        Optional[Fqcn] )
-  :   Path on Linux raises ZipError raises PathError raises IoError raises StreamError =
-
-    val entries = Classpath.Directory(compilation.out) :: compilation.classpath.entries.to(List)
-    assemble(LocalClasspath(entries*), jarfile0.or(compilation.out.peer("tmpfile.jar")), main)
-
-
-  // Bundles a compilation's output—classfiles, and any `.sjsir` or `.nir` alongside them—as a
-  // library JAR for downstream assembly: the JAR is a valid classpath entry for further
-  // compilations in the same universe, and for a later `Linker`.
-  def library[universe <: Universe]
-    ( compilation: Compilation[universe], jarfile0: Optional[Path on Linux] )
-  :   Path on Linux raises ZipError raises PathError raises IoError raises StreamError =
-
-    val entries = List(Classpath.Directory(compilation.out))
-    assemble(LocalClasspath(entries*), jarfile0.or(compilation.out.peer("tmpfile.jar")), Unset)
-
-
-  private def assemble
-    ( classpath: LocalClasspath, jarfile: Path on Linux, main: Optional[Fqcn] )
-  :   Path on Linux raises ZipError raises PathError raises IoError raises StreamError =
-
-    val manifest =
-      main.let(MainClass(_)).let: main =>
-        Manifest(ManifestVersion(()), CreatedBy(t"Soundness"), main)
-
-      . or:
-          Manifest(ManifestVersion(()), CreatedBy(t"Soundness"))
-
-
-    val omissions: Set[Text] = Set("MANIFEST.MF", "plugin.properties")
-
-    Zipfile.write(jarfile):
-      val entries =
-        Zip.Entry(%.on[Zip] / "META-INF" / "MANIFEST.MF", manifest) ::
-          classpath.entries.to(List).flatMap:
-          case ClasspathEntry.Directory(directory) =>
-            val root = directory.as[Path on Linux]
-            root.descendants.to(List).filter: entry => !omissions(entry.name)
-            . map: file =>
-              if file.entry() == Directory then Unset else
-                val ref = %.on[Zip] + root.toward(file).on[Zip]
-                Zip.Entry(ref, file.read[Data])
-
-            . compact
-
-          case ClasspathEntry.Jar(jar) =>
-            val jarfile = workingDirectory[Path on Linux].resolve(jar)
-
-            // Re-emit each entry verbatim: it already carries its compressed bytes, so no
-            // decompression or recompression is needed.
-            Zipfile.read(jarfile).entries.to(List).filter: entry =>
-              val name: Text = entry.ref.encode
-              !entry.directory && name != t"META-INF/MANIFEST.MF"
-
-          case _ =>
-            Nil
-
-      entries.distinctBy(_.ref)
-
-    jarfile
+object jarOptions:
+  // The filename of the JAR within the link's output directory.
+  def name(name: Text): Linker.Option[Artifact.Packaged] =
+    Linker.Option((_: Text) => name)
